Added list of modes to documentation, and moved colormap mode to mouse_mode.py
